어둠의 정령 이름 Gloom으로 변경, 곰 랜덤 스테이트 수정

+++ Assets/Game/02.Scripts/Boss/BossState.cs
@@ -850,7 +850,7 @@ public class BearState_Concentrate : BearState
         sphereTransform.gameObject.SetActive(false);
         helper.EndCheck();
 
-        bearController.ChangeState(eBearState.Powerless);
+        bearController.ChangeState((int)eBearState.Powerless);
     }
 }
 public class BearState_Powerless : BearState
